+++ demos/src/demos/news-feed/news-feed-detail.tsx
@@ -34,7 +34,9 @@ const DetailPage = ({ title }: { title: string }) => {
   }, []);
 
   useEffect(() => {
-    window.scrollTo(0, 0);
+    setTimeout(() => {
+      window.scrollTo(0, 0);
+    }, 0);
   }, []);
 
   const handleLike = () => {